requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
prison==0.2.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
urllib3==2.4.0
